fix compiler warnings about narrowing conversions.
(cherry picked from commit fef84f2484baa4d752d1e822f042799af2f4e117)

+++ test/write_at.cpp
@@ -96,7 +96,7 @@ public:
       const Const_Buffers& buffers)
   {
     return boost::asio::buffer_copy(
-        boost::asio::buffer(data_, length_) + offset,
+        boost::asio::buffer(data_, length_) + static_cast<size_t>(offset),
         buffers, next_write_length_);
   }
